LPD-75170 refactor: move all left files from ai-hub-rest-test to ai-hub-nexus-rest-test
