+++ modules/dxp/apps/ai-hub-nexus/ai-hub-nexus-rest-test/build.gradle
@@ -1,0 +1,30 @@
+dependencies {
+	testIntegrationImplementation group: "com.fasterxml.jackson.core", name: "jackson-annotations", version: "2.18.2"
+	testIntegrationImplementation group: "com.fasterxml.jackson.core", name: "jackson-core", version: "2.18.2"
+	testIntegrationImplementation group: "com.fasterxml.jackson.core", name: "jackson-databind", version: "2.18.2"
+	testIntegrationImplementation group: "com.liferay", name: "jakarta.ws.rs", version: "3.1.0.LIFERAY-PATCHED-1"
+	testIntegrationImplementation group: "commons-beanutils", name: "commons-beanutils", version: "1.11.0"
+	testIntegrationImplementation group: "jakarta.annotation", name: "jakarta.annotation-api", version: "2.1.1"
+	testIntegrationImplementation group: "org.skyscreamer", name: "jsonassert", version: "1.2.3"
+	testIntegrationImplementation project(":apps:account:account-api")
+	testIntegrationImplementation project(":apps:ai-hub-cell:ai-hub-cell-api")
+	testIntegrationImplementation project(":apps:headless:headless-batch-engine:headless-batch-engine-client")
+	testIntegrationImplementation project(":apps:oauth2-provider:oauth2-provider-api")
+	testIntegrationImplementation project(":apps:oauth2-provider:oauth2-provider-scope-api")
+	testIntegrationImplementation project(":apps:object:object-api")
+	testIntegrationImplementation project(":apps:object:object-rest-api")
+	testIntegrationImplementation project(":apps:object:object-test-util")
+	testIntegrationImplementation project(":apps:portal-configuration:portal-configuration-test-util")
+	testIntegrationImplementation project(":apps:portal-odata:portal-odata-api")
+	testIntegrationImplementation project(":apps:portal-search:portal-search-test-util")
+	testIntegrationImplementation project(":apps:portal-vulcan:portal-vulcan-api")
+	testIntegrationImplementation project(":apps:portal-workflow:portal-workflow-api")
+	testIntegrationImplementation project(":apps:portal-workflow:portal-workflow-kaleo-api")
+	testIntegrationImplementation project(":apps:portal-workflow:portal-workflow-kaleo-definition-api")
+	testIntegrationImplementation project(":apps:portal-workflow:portal-workflow-kaleo-runtime-api")
+	testIntegrationImplementation project(":apps:site:site-api")
+	testIntegrationImplementation project(":dxp:apps:ai-hub-nexus:ai-hub-nexus-api")
+	testIntegrationImplementation project(":dxp:apps:ai-hub-nexus:ai-hub-nexus-rest-api")
+	testIntegrationImplementation project(":dxp:apps:ai-hub-nexus:ai-hub-nexus-rest-client")
+	testIntegrationImplementation project(":test:arquillian-extension-junit-bridge")
+}
