+++ modules/dxp/apps/ai-hub-nexus/ai-hub-nexus-rest-test/src/testIntegration/resources/com/liferay/ai/hub/nexus/rest/resource/v1_0/test/dependencies/ai-decision-node-with-tool-workflow-definition.json
